Make html reporter preserve order of benchmark runs
Benchmarks are run in the order they are written in the file.  If the
reporter preserves it, this gives the benchmarks author control over
what is the best way to present the results of the benchmarks.

However, this was not the case because benchmark results where stored in
an `unsorted_map` that has an unspecified order.  By storing them in a
vector that is populated as the benchmarks are run, we do preserve the
run order.